+++ controllers/Types.js
@@ -1,5 +1,5 @@
-const Types = require('../models/Types'); // Adjust the path as necessary
-
+const Types = require('../models/Types'); 
+const Tickets = require('../models/Tickets');
 // Create a new type
 exports.createType = async (req, res) => {
     try {
@@ -56,16 +56,43 @@ exports.updateType = async (req, res) => {
 exports.deleteType = async (req, res) => {
     try {
         const { id } = req.params;
+
+        // Find a ticket with the type_id that is being deleted
+        const ticket = await Tickets.findOne({ where: { type_id: id } });
+
+        if (!ticket) {
+            return res.status(404).json({ error: 'No tickets found with the given type.' });
+        }
+
+        // Get the project GUID from the ticket
+        const ticketProjGuid = ticket.project_guid;
+
+        // Find the default type for the project
+        const defaultTypeDetails = await Types.findOne({ where: { project_guid: ticketProjGuid, isDefault: true } });
+
+        if (!defaultTypeDetails) {
+            return res.status(404).json({ error: 'Default type not found for this project.' });
+        }
+
+        const defaultTypeId = defaultTypeDetails.id;
+
+        // Update tickets to assign the default type where the old type was used
+        await Tickets.update({ type_id: defaultTypeId }, {
+            where: { type_id: id }
+        });
+
+        // Delete the old type
         const deleted = await Types.destroy({
             where: { id }
         });
+
         if (deleted) {
-            res.status(204).send(); // No content
+            res.status(200).send('Type successfully deleted and tickets reassigned to default type.');
         } else {
             res.status(404).json({ error: 'Type not found.' });
         }
     } catch (error) {
-        res.status(500).json({ error: 'An error occurred while deleting the type.' });
+        res.status(500).json({ error: 'An error occurred while deleting the type.', details: error.message });
     }
 };
